removed comments & added CacheProvider
