+++ TheaterDatabase/dbo/Tables/Wishlist.sql
@@ -3,7 +3,6 @@
     [AccountId]         INT NOT NULL,
     [PerformanceId]     INT NOT NULL,
     CONSTRAINT [PK_Wishlist] PRIMARY KEY CLUSTERED ([WishPerformanceId] ASC),
-    CONSTRAINT [FK_Wishlist_Account] FOREIGN KEY ([AccountId]) REFERENCES [dbo].[Account] ([AccountId]),
-    --CONSTRAINT [FK_Wishlist_Performance] FOREIGN KEY ([PerformanceId]) REFERENCES [dbo].[Performance] ([PerformanceId])
+    CONSTRAINT [FK_Wishlist_Account] FOREIGN KEY ([AccountId]) REFERENCES [dbo].[Account] ([AccountId])
 );
