+++ TheaterSchedule.BLL/DTO/WishlistDTO.cs
@@ -4,6 +4,6 @@
     {
         public int PerformanceId { get; set; }
         public string Title { get; set; }
-        public /*byte[]*/ string MainImage { get; set; }
+        public string MainImage { get; set; }
     }
 }
